Move bmath test info
